Fixes Test Package Declarations
