+++ src/test/java/org/edgexfoundry/support/notifications/dao/integration/mongodb/TransmissionDAOTest.java
@@ -27,8 +27,8 @@ import java.util.List;
 import org.edgexfoundry.EdgeXSupportNotificationsApplication;
 import org.edgexfoundry.support.domain.notifications.Notification;
 import org.edgexfoundry.support.domain.notifications.Transmission;
-import org.edgexfoundry.support.notification.test.data.NotificationData;
-import org.edgexfoundry.support.notification.test.data.TransmissionData;
+import org.edgexfoundry.support.notifications.test.data.NotificationData;
+import org.edgexfoundry.support.notifications.test.data.TransmissionData;
 import org.edgexfoundry.support.notifications.dao.NotificationDAO;
 import org.edgexfoundry.support.notifications.dao.TransmissionDAO;
 import org.edgexfoundry.test.category.RequiresMongoDB;
