+++ src/test/java/org/edgexfoundry/support/notifications/service/NormalSeverityResendExecutorTest.java
@@ -1,0 +1,58 @@
+/*******************************************************************************
+ * Copyright 2016-2017 Dell Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except
+ * in compliance with the License. You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software distributed under the License
+ * is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express
+ * or implied. See the License for the specific language governing permissions and limitations under
+ * the License.
+ *
+ * @microservice: support-notifications
+ * @author: Jim White, Dell
+ * @version: 1.0.0
+ *******************************************************************************/
+
+package org.edgexfoundry.support.notifications.service;
+
+import org.edgexfoundry.support.notifications.config.GeneralConfig;
+import org.edgexfoundry.support.notifications.dao.TransmissionDAO;
+import org.edgexfoundry.support.notifications.service.DistributionCoordinator;
+import org.edgexfoundry.support.notifications.service.impl.NormalSeverityResendExecutor;
+import org.edgexfoundry.test.category.RequiresNone;
+import org.junit.Before;
+import org.junit.Test;
+import org.junit.experimental.categories.Category;
+import org.mockito.InjectMocks;
+import org.mockito.Mock;
+import org.mockito.MockitoAnnotations;
+
+@Category(RequiresNone.class)
+public class NormalSeverityResendExecutorTest {
+
+  @InjectMocks
+  private NormalSeverityResendExecutor executor;
+
+  @Mock
+  private TransmissionDAO dao;
+
+  @Mock
+  private DistributionCoordinator distributionCoordinator;
+
+  @Mock
+  private GeneralConfig generalConfig;
+
+  @Before
+  public void setup() {
+    MockitoAnnotations.initMocks(this);
+  }
+
+  @Test
+  public void testStartResending() {
+    executor.startResnding();
+  }
+
+}
